Fix failure to parse IPv6 fc00::9bf:aca2:5b82:bbe4
The code generated from the ABNF cannot handle this.

After checking that this address complies with the ABNF in RFC5954
(https://tools.ietf.org/html/rfc5954) and that the IPv6 part of the
ABNF in src/org/murillo/abnf/sdp.abnf matches the one in the RFC
I turned eye to the generator.

The code was generated with aparse-2.2.
Regenerating with aparse-2.5 (and changing invocation sites due to
naming pattern change between 2.2 and 2.5) the problem was solved.